cleanup commands structure
